use case simpification #219
